+++ mio/process/video.py
@@ -8,7 +8,7 @@ from pathlib import Path
 import cv2
 import numpy as np
 import pandas as pd
-from tqdm import tqdm
+from tqdm import tqdm, trange
 
 from mio import init_logger
 from mio.io import VideoReader, VideoWriter
@@ -21,7 +21,11 @@ from mio.models.process import (
     NoisePatchConfig,
 )
 from mio.plots.video import VideoPlotter
-from mio.process.frame_helper import FrequencyMaskHelper, InvalidFrameDetector
+from mio.process.frame_helper import (
+    FrequencyMaskHelper,
+    InvalidFrameDetector,
+    make_detectors,
+)
 from mio.process.zstack_helper import ZStackHelper
 
 logger = init_logger("video")
@@ -131,15 +135,9 @@ class NoisePatchProcessor(BaseVideoProcessor):
         self.noise_detect_helper = InvalidFrameDetector(noise_patch_config=noise_patch_config)
         self.noise_patchs: list[np.ndarray] = []
         self.noisy_frames: list[np.ndarray] = []
-        self.diff_frames: list[np.ndarray] = []
         self.dropped_frame_indices: list[int] = []
 
         self.output_enable: bool = noise_patch_config.output_result
-
-        if "mean_error" in noise_patch_config.method:
-            logger.warning(
-                "The mean_error method is unstable and not fully tested yet." " Use with caution."
-            )
 
     def process_frame(self, input_frame: np.ndarray, index: int) -> np.ndarray | None:
         """
@@ -186,15 +184,6 @@ class NoisePatchProcessor(BaseVideoProcessor):
         return NamedVideo(name="patched_area", video=self.noise_patchs)
 
     @property
-    def diff_frames_named_video(self) -> NamedVideo:
-        """
-        Get the NamedFrame object for the difference frames.
-        """
-        if not hasattr(self.noise_patch_config, "diff_multiply"):
-            diff_multiply = 1
-        return NamedVideo(name=f"diff_{diff_multiply}x", video=self.diff_frames)
-
-    @property
     def noisy_frames_named_video(self) -> NamedVideo:
         """
         Get the NamedFrame object for the noisy frames.
@@ -217,18 +206,6 @@ class NoisePatchProcessor(BaseVideoProcessor):
         else:
             logger.info(f"{self.name} noise patch output disabled.")
 
-    def export_diff_frames(self) -> None:
-        """
-        Export the difference frames to a file.
-        """
-        if self.noise_patch_config.output_diff:
-            logger.info(f"Exporting {self.name} difference frames to {self.output_dir}")
-            self.diff_frames_named_video.export(
-                output_path=self.output_dir / f"{self.name}", fps=20, suffix=True, force=self.force
-            )
-        else:
-            logger.info(f"{self.name} difference frames output disabled.")
-
     def export_noisy_video(self) -> None:
         """
         Export the noisy frames to a file.
@@ -250,7 +227,6 @@ class NoisePatchProcessor(BaseVideoProcessor):
         """
         self.export_output_video()
         self.export_noise_patch()
-        self.export_diff_frames()
         self.export_noisy_video()
 
 
@@ -583,7 +559,6 @@ def denoise(
 
     noise_patch_processor.output_dir = debug_dir
     noise_patch_processor.export_noise_patch()
-    noise_patch_processor.export_diff_frames()
     noise_patch_processor.export_noisy_video()
 
     freq_mask_processor.batch_export_videos()
@@ -633,6 +608,44 @@ def denoise(
         video_plotter.show()
 
     return output_video_path
+
+
+def score_noise(
+    recording: Recording, config: NoisePatchConfig, progress: bool = False
+) -> pd.DataFrame:
+    """
+    Score framewise noise from a recording,
+    yielding a dataframe with columns for each kind of noise
+
+    - reconstructed_frame_index: the index of the frame in the video
+    - gradient: the number of pixels that are part of noise patches,
+      as determined by the second row-wise derivative being above the configured threshold
+    - black_area: the number of pixels in contiguous black regions (and thus missing)
+    """
+
+    records = []
+    detectors = make_detectors(config)
+    n_frames = recording.video.n_frames
+    iterator = trange(n_frames) if progress else range(n_frames)
+
+    try:
+        for idx in iterator:
+            record = {"reconstructed_frame_index": idx}
+            frame = recording.video[idx]
+            # FIXME: video proxy should return grayscale as grayscale
+            # https://github.com/miniscope/mio/issues/175
+            if len(frame.shape) == 3 and frame.shape[-1] == 3:
+                frame = cv2.cvtColor(frame, cv2.COLOR_BGR2GRAY)
+            for method, detector in detectors.items():
+                _, noise_mask = detector.find_invalid_area(frame)
+                record[method] = np.count_nonzero(noise_mask)
+            records.append(record)
+
+    finally:
+        if progress:
+            iterator.close()
+
+    return pd.DataFrame(records)
 
 
 def trim(