+++ tests/test_process/test_stitch.py
@@ -17,7 +17,10 @@ from mio.models.dataset import Recording, Dataset, StitchedRecording
 from mio.process.stitch import (
     CandidateFrame,
     StitchRecord,
+    concat_recordings,
     stitch,
+    _align_by_time,
+    _has_discontinuous_runs,
     _score_edges,
 )
 from mio.process.video import trim, remove_frames
@@ -25,7 +28,7 @@ from mio.utils import hash_video
 
 STITCH_DATA_DIR = Path(__file__).parent.parent / "data" / "stitch"
 
-EXPECTED_STITCHED_VIDEO_HASH = "c8cdf3149f812ae25e6f3f1a876249e4ce118e9a53aa1805e48b995b01f07a91"
+EXPECTED_STITCHED_VIDEO_HASH = "df937c8651cf142b4d8e2a75140729dcacdc1151ebc3767b48d0ca71578007ff"
 EXPECTED_DEBUG_VIDEO_HASH = (
     "856e6e5c538532bd0fcfb942616686a5cd262aadb51dd8796adf5de69215c94b",
     "a69b6cadf4ab1dd8a1097d2c1be298397206db235fd4c5f68febd1700f15a4b6",
@@ -119,8 +122,10 @@ def test_score_csv_edge_scoring_tiebreaker(stitch_result: StitchedRecording):
     df = stitch_result.scores
     # filter frames where only one video or the other had them
     df = df[~df["compare_video"].isna()]
-    # there should be four frames that could be decided on metadata alone
-    assert len(df[df["selected_edge_score"].isna()]) == 4
+    # there should be 7 frames that could be decided on metadata alone
+    # - 4x on buffer count
+    # - 3x on black pixels
+    assert len(df[df["selected_edge_score"].isna()]) == 7
     # for all those that had to use edge scores, the selected should be greater or equal
     edges_scored = df[~df["selected_edge_score"].isna()]
     for _, row in edges_scored.iterrows():
@@ -287,3 +292,123 @@ def test_remove_frames_invalid(tmp_path):
             remove_indices=list(range(EXPECTED_VIDEO1_FRAME_COUNT)),
             output_path=tmp_path / "out.avi",
         )
+
+
+def test_concat_recordings(tmp_path, recordings):
+    """Concatenating two recordings produces contiguous frame indices and correct frame count."""
+
+    combined_video = tmp_path / "combined.avi"
+
+    combined = concat_recordings(
+        recordings=list(recordings.values()),
+        output_video_path=combined_video,
+    )
+
+    # Video frame count should be sum of both inputs
+    expected_frames = sum(r.video.n_frames for r in recordings.values())
+    # the Recording class also validates that the metadata has a matching length if present
+    # so its presence means that the metadata is also matching
+    assert combined.metadata is not None
+
+    actual_frames = combined.video.n_frames
+    assert actual_frames == expected_frames
+
+    # CSV should have contiguous reconstructed_frame_index
+    df = combined.metadata
+    diffs = df["reconstructed_frame_index"].diff().iloc[1:].to_numpy()
+    assert (diffs <= 1).all() and (diffs >= 0).all()
+
+
+@pytest.mark.parametrize("flip", [True, False])
+def test_align_by_time(tmp_path, flip):
+    """Aligning by timestamp finds the inner join of the closest matching timestamps"""
+    # one normal video with some linspaced times
+    left_idxes = np.ravel(np.repeat(np.arange(50), 5))
+    left_times = np.linspace(0, 1, len(left_idxes))
+    left = pd.DataFrame(
+        {"reconstructed_frame_index": left_idxes, "buffer_recv_unix_time": left_times}
+    )
+
+    # one offset video with a blippy frame from a bit flip in the frame_num
+    right_idxes = np.ravel(np.repeat(np.arange(25), 5))
+    right_idxes = np.concat([right_idxes, [25]], axis=0)
+    right_idxes = np.concat([right_idxes, np.ravel(np.repeat(np.arange(26, 52), 5))], axis=0)
+    # make same size so sampling rate is the same
+    right_idxes = right_idxes[: len(left_idxes)]
+    right_times = np.linspace(0, 1, len(right_idxes)) + 0.1
+    right = pd.DataFrame(
+        {"reconstructed_frame_index": right_idxes, "buffer_recv_unix_time": right_times}
+    )
+
+    good, bad = "video1", "video2"
+    if flip:
+        good, bad = "video2", "video1"
+    recordings = [
+        Recording.model_construct(name=good, metadata=left),
+        Recording.model_construct(name=bad, metadata=right),
+    ]
+
+    aligned = _align_by_time(recordings)
+    # we should have received 45 frames: 50 frames in the original - 5 frames in 0.1 seconds of lag
+    assert len(aligned) == 45
+    assert np.array_equal(aligned[good], np.arange(5, 50))
+
+    # we should have dropped frame 25 in the right one
+    assert 25 not in np.array(aligned[bad])
+    assert np.array_equal(aligned[bad], np.concat([np.arange(25), np.arange(26, 46)]))
+
+
+def test_stitch_with_timestamps(stitch_result, tmp_path):
+    """
+    When we scramble the `frame_num`, we can stitch by timestamps.
+    We should get the same result as if we were able to use frame_num in this case.
+    """
+    # use a temporary version of the recordings because we are going to wreck the metadata
+    recordings = {
+        "video1": Recording.from_video(STITCH_DATA_DIR / "video1.avi"),
+        "video2": Recording.from_video(STITCH_DATA_DIR / "video2.avi"),
+    }
+    recordings["video1"].metadata["frame_num"] = np.random.default_rng().integers(
+        0, 1000, size=len(recordings["video1"].metadata)
+    )
+    recordings["video2"].metadata["frame_num"] = np.random.default_rng().integers(
+        0, 1000, size=len(recordings["video2"].metadata)
+    )
+
+    result = stitch(list(recordings.values()), debug_video=True, output_dir=tmp_path)
+    # we should have an inner join on the frames - so only those without a comparison frame
+    expected = stitch_result.scores[~stitch_result.scores["compare_video"].isna()]
+    assert np.array_equal(result.scores["selected_video"], expected["selected_video"])
+
+
+@pytest.mark.parametrize(
+    "series,expected",
+    [
+        pytest.param([1, 1, 1, 2, 2, 2, 3, 3, 3], False, id="contiguous-buffers"),
+        pytest.param([1, 2, 3, 4, 5], False, id="contiguous-frames"),
+        pytest.param([1, 1, 1, 2, 500, 2, 3, 3, 3], False, id="single-bitflip-same-frame"),
+        pytest.param([1, 1, 1, 2, 2, 500, 3, 3, 3], False, id="single-bitflip-next-frame"),
+        pytest.param(
+            [10, 10, 10, 11, 11, 11, 2, 2, 2, 3, 3, 3], True, id="discontiguous-buffers-lower"
+        ),
+        pytest.param(
+            [10, 10, 10, 11, 11, 11, 20, 20, 20, 21, 21, 21],
+            True,
+            id="discontiguous-buffers-higher",
+        ),
+        pytest.param([1, 2, 3, 4, 5, 1, 2, 3, 4, 5], True, id="discontiguous-frames-lower"),
+        pytest.param([1, 2, 3, 4, 5, 10, 11, 12, 13], True, id="discontiguous-frames-higher"),
+    ],
+)
+def test_has_discontinuous_runs(series, expected):
+    """
+    We can determine when some timeseries has discontinuous runs,
+    ignoring when a single value blips incorrectly (like a bit flip in a metadata header).
+    """
+    series = pd.Series(series)
+    assert _has_discontinuous_runs(series) == expected
+
+
+def test_test_data_is_considered_continuous(recordings):
+    """Just testing the assumptions of the tests ios all"""
+    assert not any(_has_discontinuous_runs(r.metadata["frame_num"]) for r in recordings.values())