+++ tests/test_process/test_video.py
@@ -46,18 +46,6 @@ def test_noise_patch_processor(video_frame, tmp_path):
     assert processor.output_enable
 
 
-def test_noise_patch_processor_no_config(random_8bit_video_frame, tmp_path):
-    denoise_config = DenoiseConfig.from_id("denoise_example")
-    denoise_config.noise_patch.enable = True
-    denoise_config.noise_patch.mean_error_config = None
-    denoise_config.noise_patch.gradient_config = None
-    denoise_config.noise_patch.black_area_config = None
-
-    # This should raise a ValueError because the necessary configs are not provided
-    with pytest.raises(ValueError):
-        NoisePatchProcessor("denoise_example", denoise_config.noise_patch, tmp_path)
-
-
 def test_noise_patch_processor_no_methods(random_8bit_video_frame, tmp_path):
     denoise_config = DenoiseConfig.from_id("denoise_example")
     denoise_config.noise_patch.enable = True